requirements.txt
Flask
Flask_SQLAlchemy
python-dotenv
psycopg2-binary
authlib
requests
